# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.StatusResponse:
    properties:
      status:
        example: deleted
        type: string
    type: object
...
  handlers.WorkoutRequest:
    properties:
      avg_cadence_spm:
        type: integer
      avg_hr:
        type: integer
      avg_speed_cm_s:
        type: integer
      calories:
        type: number
      distance_m:
        type: integer
      downhill_cm:
        type: integer
      duration_s:
        type: integer
      elevation_cm:
        type: integer
      end_utc:
        description: RFC3339 optional
        type: string
      locations:
        items:
          $ref: '#/definitions/store.WorkoutLocation'
        type: array
      max_hr:
        type: integer
      max_speed_cm_s:
        type: integer
      min_hr:
        type: integer
      source:
        type: string
      sport_count:
        type: integer
      sport_type_id:
        type: integer
      start_utc:
        description: RFC3339
        type: string
      steps:
        type: integer
      tz_offset_min:
        type: integer
      uphill_cm:
        type: integer
    type: object
  httpx.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  store.WorkoutLocation:
    properties:
      rate_real:
        type: integer
    type: object
  store.WorkoutResponse:
    properties:
      avg_cadence_spm:
        type: integer
      avg_hr:
        type: integer
      avg_speed_cm_s:
        type: integer
      calories:
        type: number
      device_id:
        type: string
      distance_m:
        type: integer
      downhill_cm:
        type: integer
      duration_s:
        type: integer
      elevation_cm:
        type: integer
      end_utc:
        type: string
      locations:
        items:
          $ref: '#/definitions/store.WorkoutLocation'
        type: array
      max_hr:
        type: integer
      max_speed_cm_s:
        type: integer
      min_hr:
        type: integer
      source:
        type: string
      sport_count:
        type: integer
      sport_type_id:
        type: integer
      start_utc:
        type: string
      steps:
        type: integer
      synced_at:
        type: string
      tz_offset_min:
        type: integer
      uphill_cm:
        type: integer
    type: object
info:
  contact: {}
  description: Backend API for AGEM wearable users, devices, qring sensor data, sleep,
    and workouts.
  title: AGEM Backend API
  version: "1.0"
paths:
...
  /v1/devices/{device_id}/workouts:
    delete:
      description: Deletes a workout identified by device ID and start_utc query parameter.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Workout start timestamp
        format: date-time
        in: query
        name: start_utc
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Delete workout
      tags:
      - Workouts
    get:
      description: Lists workouts in an RFC3339 time range.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Inclusive RFC3339 start timestamp
        format: date-time
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive RFC3339 end timestamp
        format: date-time
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.WorkoutResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List workouts
      tags:
      - Workouts
    patch:
      consumes:
      - application/json
      description: Updates a workout identified by device ID and start_utc in the
        request body.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Workout update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WorkoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.WorkoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Update workout
      tags:
      - Workouts
    post:
      consumes:
      - application/json
      description: Creates or replaces a workout keyed by device ID and start_utc.
      parameters:
      - description: Device ID
        in: path
        name: device_id
        required: true
        type: string
      - description: Workout payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.WorkoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.WorkoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Create workout
      tags:
      - Workouts
...
swagger: "2.0"